pyproject.toml
[tool.poetry]
name = "streamlabsio"
version = "0.1.0"
description = "Get real time Twitch/Youtube events through Streamlabs SocketIO API"
authors = ["onyx-and-iris <code@onyxandiris.online>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
tomli = { version = "^2.0.1", python = "<3.11" }
python-engineio = "3.14.2"
python-socketio = {version = "4.6.0", extras = ["client"]}
observable = "^1.0.3"

[tool.poetry.dev-dependencies]
black = "^22.10.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
